% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-code.R
\name{logliknormLR}
\alias{logliknormLR}
\title{Fast Computation of Marginal Predictive Loglikelihood}
\usage{
logliknormLR(yy, XX_aug, alpha, init)
}
\arguments{
\item{yy}{T-length time series vector.  y[1] represents the beginning of
the time eries.}

\item{XX_aug}{[T x (p+1)] dimensional matrix of covariates. This will
include the intercept column.}

\item{alpha}{PWD parameter we are calculating the marginal predictive
loglikelihood for.}

\item{init}{integer representing the time point to begin computing marginal
predictive loglikelihood.}
}
\value{
Return a scalar value representing the marginal predictive 
loglikelihood of the data given alpha.
}
\description{
Helper function which takes as input a vector-valued response, y, 
a predictor matrix, X, a particular value of alpha, and init. It
returns as an output the marginal predictive loglikelihood of the 
data given that value of alpha.
}

