\name{PWD-package}
\alias{PWD-package}
\alias{PWD}
\docType{package}
\title{
Time Series Regression Using the Power Weighted Densities (PWD) Approach
}
\description{
  
Contains functions which allow the user to perform time series 
regression quickly using the Power Weighted Densities (PWD) approach. 
alphahat_LR_one_Rcpp() is the main workhorse function within this package.

}
\details{
  \tabular{ll}{
    Package: \tab PWD\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2016-02-27\cr
    License: \tab GPL-3\cr
    }

}
\author{
  Daniel McCarthy

  Maintainer: Daniel McCarthy <danielm@wharton.upenn.edu>
}
\references{

McCarthy, Daniel; Jensen, Shane. "Power Weighted Densities for Time Series Data." December 2015. Available at arXiv: http://arxiv.org/abs/1412.4059

}

\keyword{ package }

